requirements.txt
streamlit
faker
pandas
numpy
openpyxl
python-dateutil
pytest
